% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_cols.R
\name{arrange_cols}
\alias{arrange_cols}
\title{Arrange columns}
\usage{
arrange_cols(dt, at, ...)
}
\arguments{
\item{dt}{input data.table}

\item{at}{position for inserting the selected columns.  This can be any valid
column names or "start" or "end" for placing the selected column before the
first column or after the last column}

\item{...}{a vector of characters of column names or column names seperated by commas (NSE)}
}
\description{
This wrapper function make use of the setcolorder() function and
arrange columns by place selected columns after the specified position.
}
