% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_rows.R
\name{ascend_rows}
\alias{ascend_rows}
\title{Arranging rows by selected variables}
\usage{
ascend_rows(dt, ...)
}
\arguments{
\item{dt}{input data.table}

\item{...}{column names or a vector of characters of column names in the data.table to be used for ordering
the raws of data}
}
\description{
Arranging rows of data in the data.table by selected variables
in ascending order. This is similar to the arrange() function in the dplyr
package.
}
\examples{
\dontrun{
 # in descending order by values in weight and mpg
 dt <- data1 |> data.table()

 dt |> ascend_rows(wt, mpg)

 # group using cyl and order by wt and mpg
 dt |> set_group(cyl) |> ascend_rows(wt, mpg)
}
}
