% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_helpers.R
\name{char_to_symbol}
\alias{char_to_symbol}
\title{Helper function to turn characters to symbols}
\usage{
char_to_symbol(characters)
}
\arguments{
\item{characters}{characters or vector of characters to be turned into symbol}
}
\description{
useful when programming with functions in this package, when
variable names used to store column names as characters are used as
arugments. It is simply a wrapper function for the case of using the
parse() function that convert characters into an unevaluated symbol
}
