% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_withRownames.R
\name{data.table.withRownames}
\alias{data.table.withRownames}
\title{Create data.table from data.frame with row names}
\usage{
data.table.withRownames(df, check.names=FALSE, key = NULL, stringsAsFactors = FALSE)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{check.names}{same as in data.frame}

\item{key}{character vector of one or more column names which is passed to setkey}

\item{stringsAsFactors}{same as in data.table}
}
\value{
returns a \code{data.table}
}
\description{
This a wrapper function around \code{data.table(df, keep.rownames=TRUE)}

For more information about \code{data.table} and \code{data.frame} \href{http://www.google.com}{Google it}.
}
\examples{
data.table.withRownames(mtcars)

\dontrun{
data.table.withRownames(classesOtherThanDataFrame)
}

}
