% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Example data using mtcars}
\format{
A \code{data.frame} with columns:
\describe{
\item{mpg}{Miles per US gallon}
\item{cyl}{No. of engine cylinder}
\item{disp}{Engine displacement (cu. in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight of the vehicle lb/1000}
\item{qsec}{1/4 mile time}
\item{vs}{Engine cylinder configuration 0 = v-shape, 1 = in-line}
\item{am}{0 = automatic, 1 = manual}
\item{carb}{Number of carburetors}
}
}
\source{
from base R
}
\usage{
data1
}
\description{
mtcars - data.frame with row names
}
\examples{
\dontrun{
  data1
}
# data can be found in
system.file("data", package = "wr.data.table")
}
\keyword{datasets}
