% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{Example data using iris}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\source{
from base R
}
\usage{
data2
}
\description{
iris data
}
\examples{
\dontrun{
  data2
}
# data can be found in
system.file("data", package = "wr.data.table")
}
\keyword{datasets}
