% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{data3}
\alias{data3}
\title{Example data using cars}
\format{
A \code{data.frame} with columns:
\describe{
\item{speed}{miles per hour}
\item{dist}{distance in km}
}
}
\source{
from base R
}
\usage{
data3
}
\description{
cars data
}
\examples{
\dontrun{
  data3
}
# data can be found in
system.file("data", package = "wr.data.table")
}
\keyword{datasets}
