% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cols.R
\name{def_cols}
\alias{def_cols}
\title{Create and assign new columns in data.table}
\usage{
def_cols(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{formula using \code{~} notation where LHS is the new column name and RHS is an expression for the values
More than one new columns can be created by having formulas separated by commas}
}
\value{
returns a \code{data.table}
}
\description{
function used to create new columns in data.table using non-standard evaluation
}
\examples{
\dontrun{
dt <- data.table(mtcars)
define(dt, kmpl ~ mpg / 2.35215, wt_in_kg ~ wt / 2.20462)
# creating two columns / converted units
define(dt, am = NULL) # remove a column
}
}
