% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descend_rows.R
\name{descend_rows}
\alias{descend_rows}
\title{Arranging rows by selected variables}
\usage{
descend_rows(dt, ...)
}
\arguments{
\item{dt}{input data.table}

\item{...}{column/variable names in the data.table to be used for ordering
the raws of data}
}
\description{
Arranging rows of data in the data.table by selected variables
in descending order. This is similar to the arrange() function in the dplyr
package.
}
\examples{
\dontrun{
# in descending order by values in weight and mpg
dt |> descend_rows(wt, mpg)
# group using cyl and order by wt and mpg
dt |> set_group(cyl) |> descend_rows(wt, mpg)
}
}
