% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desel_cols.R
\name{desel_cols}
\alias{desel_cols}
\title{Function to subset data.table by removal columns/variables}
\usage{
desel_cols(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{some parameters}
}
\description{
This function is similar to
}
