% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Function to filter rows of data by condition}
\usage{
filter_rows(dt, condition)
}
\arguments{
\item{dt}{input \code{data.table}}

\item{condition}{a valid R conditional expression}
}
\description{
wrapper function around filtering data.table with given conditions with non-standard evaluation
}
\details{
This function simply takes the input conditions
}
\examples{
\dontrun{
filter_rows(data.table(mtcars),mpg > 20)  # return all rows where mpg is > 20
filter_rows(data.table(mtcars), mpg > 20, vs = 1, am = 0)
# return all rows where mpg is > 20, vs = 1, am = 0

wr.filter(data.table(mtcars), "mpg > 20") # error - conditions should not be parsed as characters
}

}
