% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdt.R
\name{mkdt}
\alias{mkdt}
\title{Function for turning mtcars data.frame into data.table}
\usage{
mkdt()
}
\description{
for convenience during debugging.  Creates a data.table called \code{dt} in the global environment
}
