% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_group.R
\name{rm_group}
\alias{rm_group}
\title{Ungroup data}
\usage{
rm_group(dt)
}
\arguments{
\item{dt}{a \code{data.table}}
}
\description{
this function simply removes the $group in the \code{data.table} attributes
}
