% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn_cols.R
\name{rn_cols}
\alias{rn_cols}
\title{Setting column names}
\usage{
rn_cols(dt, old=NULL, new=NULL, ...)
}
\arguments{
\item{dt}{input data.table}

\item{old}{default = NULL vector of existing column names}

\item{new}{default = NULL vector of new column names}

\item{...}{non-standard evaluation of in the form of an expression such as,
old_column_name = new_column_name}
}
\description{
using data.table packages setnames function to change column
names
}
\details{
User can use provide a vector of characters for current column names
and a vector of characters of equal lengths as new column names.
Alternatively, user can also use the old_column_name=new_column_name
notation, separated by comma.
}
\examples{
\dontrun{
rn_cols(dt, hp=horsepower, disp=displacement)
rn_cols(dt, old=c("hp", "disp", "cyl"), new=c("horsepower", "displacement", "cylinder"))
}
}
