% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_group.R
\name{set_group}
\alias{set_group}
\title{Group data using variables}
\usage{
set_group(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{names of variables that are used for grouping}
}
\description{
add $group attribute to data.table
$group is a vector of variable anmes
}
