% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{Long to short data}
\usage{
spread(dt, key, value)
}
\arguments{
\item{dt}{input data.table}

\item{key}{key variable used to spread data}

\item{value}{data values for new columns}
}
\description{
Long to short data
}
