% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr.group_by.R
\name{wr.group_by}
\alias{wr.group_by}
\title{Function to group data by one or more specified variables}
\usage{
wr.group_by(dt, ...)
}
\description{
Function to group data by one or more specified variables
}
