% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr.order.R
\name{wr.order}
\alias{wr.order}
\title{Fast row reordering using data.table's order()/setorder function}
\usage{
wr.order(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{column names without quotes}
}
\value{
returns a \code{data.table}
}
\description{
This a wrapper function around data.table's row reordering function

For more information about \code{data.table} and \code{setorder()} \href{http://www.google.com}{Google it}.
}
\examples{
\dontrun{
  dt <- data.table(mtcars)
  wr.order(dt, gear, am, vs)
}
}
