% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr.select.R
\name{wr.select}
\alias{wr.select}
\title{Function to subset data.table with selected columns/variables}
\usage{
wr.select(dt, ...)
}
\description{
Function to subset data.table with selected columns/variables
}
