% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_deselect.R
\name{wr_deselect}
\alias{wr_deselect}
\title{Function to subset data.table by removal columns/variables}
\usage{
wr_deselect(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{some parameters}
}
\description{
This function is similar to
}
