% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_dt_withrownames.R
\name{wr_dt_withrownames}
\alias{wr_dt_withrownames}
\title{Create data.table from data.frame with row names}
\usage{
wr_dt_withrownames(df, check.names=FALSE, key = NULL, stringsAsFactors = FALSE)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{check.names}{same as in data.frame}

\item{key}{character vector of one or more column names which is passed to setkey}

\item{stringsAsFactors}{same as in data.table}
}
\value{
returns a \code{data.table}
}
\description{
This a wrapper function around \code{data.table(df, keep.rownames=TRUE)}

For more information about \code{data.table} and \code{data.frame} \href{http://www.google.com}{Google it}.
}
\examples{
wr_dt_withrownames(mtcars)

\dontrun{
wr_dt_withrownames(data.table(mtcars)) # this will throw an error because input is not a data.frame
df <- mtcars
rownames(df) <- NULL
wr_dt_withrownames(df)
# if data.frame has no row names, the returned data.table will create column rn
with row index numbers
}

}
