% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_groupby.R
\name{wr_groupby}
\alias{wr_groupby}
\title{Group data using variables}
\usage{
wr_groupby(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{names of variables that are used for grouping}
}
\description{
add $group attribute to data.table
$group is a vector of variable anmes
}
