% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_order.R
\name{wr_order}
\alias{wr_order}
\title{Fast row reordering using data.table's order()/setorder function}
\usage{
wr_order(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{column names.  Multiple column names are allowed.
The function will order the row by the columns in the order they are provided in the arguments}
}
\value{
returns a \code{data.table}
}
\description{
This a wrapper function around data.table's row reordering function

For more information about \code{data.table} and \code{setorder()} \href{http://www.google.com}{Google it}.
}
\examples{
\dontrun{
  dt <- data.table(mtcars)
  wr_order(dt, gear, am, vs)
}
}
