% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_renameCols.R
\name{wr_renameCols}
\alias{wr_renameCols}
\title{Setting column names}
\usage{
wr_renameCols(dt, old=NULL, new=NULL, ...)
}
\description{
using data.table packages setnames function to change column
names
}
\details{
User can use provide a vector of characters for current column names
and a vector of characters of equal lengths as new column names.
Alternatively, user can also use the old_column_name=new_column_name
notation, separated by comma.
}
\examples{
/dontrun{
wr_renameCols(dt, hp=horsepower, disp=displacement)
wr_renameCols(dt, old=c("hp", "disp", "cyl"), new=c("horsepower", "displacement", "cylinder"))
}
}
