% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_select.R
\name{wr_select}
\alias{wr_select}
\title{Function to subset data.table by selecting columns/variables}
\usage{
wr_select(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{some parameters}
}
\description{
This function is similar to
}
