% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_summarize.R
\name{wr_summarize}
\alias{wr_summarize}
\title{Summarize data}
\usage{
wr_summarize(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{some parameters}
}
\value{
returns a \code{data.table}
}
\description{
For more information about \code{data.table} and \code{setorder()} \href{http://www.google.com}{Google it}.
}
\examples{
\dontrun{
  dt <- data.table(mtcars)
}
}
