% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_ungroup.R
\name{wr_ungroup}
\alias{wr_ungroup}
\title{Ungroup data}
\usage{
wr_ungroup(dt)
}
\arguments{
\item{dt}{a \code{data.table}}
}
\description{
this function simply removes the $group in the \code{data.table} attributes
}
