% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_cols.R
\name{arrange_cols}
\alias{arrange_cols}
\title{Arrange columns}
\usage{
arrange_cols(dt, at, ...)
}
\arguments{
\item{dt}{input data.table}

\item{at}{position for inserting the selected columns.  This can be any valid
column names, which will place the selected columns after the "at" column.
Alternatively use "start" or "end" for placing the selected columns before
the first column or after the last column respectively}

\item{...}{a vector of characters of column names or column names seperated
by commas (NSE)}
}
\description{
This wrapper function make use of the setcolorder() function and
arrange columns by placing selected columns after the specified position.
}
\examples{
\dontrun{
arrange_cols(dt, at=`vehicle model`, qsec, vs, am)
arrange_cols(dt, at="start", c("qsec", "vs", "am"))
arrange_cols(dt, at="mpg", qsec:am, carb)
}
}
