% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Example data using mtcars}
\format{
A \code{data.frame} with columns:
\describe{
\item{mpg}{miles per gallon}
\item{etc}{...etc}
}
}
\source{
from base R
}
\usage{
data1
}
\description{
mtcars
}
\examples{
\dontrun{
  data1
}
# data can be found in
system.file("data", package = "wr.data.table")
}
\keyword{datasets}
