% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cols.R
\name{def_cols}
\alias{def_cols}
\title{Create and assign new columns in data.table}
\usage{
def_cols(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{expression setting up a new column.  E.g. kmpl = mpg / 2.35215.
More than one new columns can be created by having expressions separated by commas}
}
\value{
returns a \code{data.table}
}
\description{
For more information about \code{data.table} and \code{setorder()} \href{http://www.google.com}{Google it}.
}
\examples{
\dontrun{
dt <- data.table(mtcars)
def_cols(dt, kmpl = mpg / 2.35215, wt_in_kg = wt / 2.20462)
# creating two columns / converted units
def_cols(dt, am = NULL) # remove a column
}
}
