% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_cols.R
\name{sel_cols}
\alias{sel_cols}
\title{Function to subset data.table by selecting columns/variables}
\usage{
sel_cols(dt, ...)
}
\arguments{
\item{dt}{a \code{data.table}}

\item{...}{some parameters}
}
\description{
This function is similar to
}
